refactor: handtracking 폴더 기능별로 세분화
