+++ src/features/handtracking/components/HandRenderer.tsx
