+++ src/features/handtracking/utils/handState.ts
@@ -2,10 +2,6 @@
     handState.ts -> 손의 상태 구분하는 함수 로직
 */
 
-// 타입 정의
-type Landmark = { x: number; y: number; z?: number };
-type HandState = 'fist' | 'open' | 'indexUp' | 'unknown';
-
 // 거리 계산
 const dist = (a: Landmark, b: Landmark): number => {
   const dx = a.x - b.x, dy = a.y - b.y;
