+++ src/features/handtracking/utils/solveProblem.ts
